pdfReader/api/requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
easyocr==1.7.2
filelock==3.20.0
fsspec==2025.10.0
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pyclipper==1.3.0.post6
pycparser==2.23
PyMuPDF==1.26.6
pypdfium2==5.0.0
python-bidi==0.6.7
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
